Removing any types. Fragmenting code in element and function files
