+++ src/main.tsx
@@ -1,7 +1,7 @@
 import { StrictMode } from 'react'
 import { createRoot } from 'react-dom/client'
 import './index.css'
-import Game from './App.tsx'
+import Game from './Game.tsx'
 
 createRoot(document.getElementById('root')!).render(
   <StrictMode>
